purple-gio: Add purple-gio.[ch] containing purple_gio_graceful_close()
In order to gracefully close Gio connections, there can't be pending
operations with input nor output stream. Having a synchronous close
function for PurpleConnection causes just calling g_io_stream_close()
in the close function to be a potential race condition.

This patch adds a function which takes a GIOStream and optionally any
wrapping GInputStream or GOutputStream, waits for pending operations
to complete asynchronously if any, and then gracefully closes the
connection. If no operations are pending, it closes them synchronously.