Add sub/libpurple/data/purple-url-handler.desktop.in to POTFILES.skip
Because of a bug in intltools, when running `make distcheck`, it
complains about an .in files which are in this sub directory.
See https://bugs.launchpad.net/intltool/+bug/1117944 for details.

This has already been addressed before, but the new
purple-url-handler.desktop.in wasn't added since this was resolved.
This patch adds it.