Move status icons to the status icon theme specification directory
This patch moves the status icons into a location which can be used
by GtkIconTheme. It also renames a few of them ('available' =>
'user-available' and similar) to more closely follow the freedesktop
icon theme specification.

'user-extended-away' and 'user-busy' aren't technically standard,
but I changed them as well to be more consistent with the pattern.
A couple more of them may be good to change, but I wasn't sure, so
they can be changed in a later commit if desired.

It also updates code which directly accesses the icon files.