libpidgin: Drop --multiple option for *nix
This patch drops the --multiple option and therefore unique instance
checking for *nix versions of Pidgin. This is done in preparation for
porting Pidgin to use GApplication's command line parsing and
instance checking (see the following commits). "--gapplication-app-id"
will be the new flag to allow multiple instances.

The Windows version uses its own, different method of checking for
uniqueness.