screencap: Fix converting GdkPixbuf to PurpleImage
This patch fixes converting a GdkPixbuf to a PurpleImage in the
Screen Capture plugin. gdk_pixbuf_save_to_callback() calls the
callback passed multiple times with chunks of the image. The replaced
implementation created a new PurpleImage for each chunk. Also, due
to a problem with the way pointers were used, it would discard each
PurpleImage, leaking them. This patch opts to instead use
gdk_pixbuf_save_to_buffer() and passes it off to
purple_image_take_data() for both simplicity and to fix the above
problems.