pidgin: Fix GApplication activation
If G_APPLICATION_HANDLES_COMMAND_LINE is passed to the GApplication,
it no longer automatically calls g_application_activate() if a
duplicate instance is run with no additional command line arguments.
This patch adds a call to g_application_activate() in the command
line callback in order to restore this functionality.