Pidgin: Drop specific Pango dependency check
We depend on Pango >= 1.4 for decent RTL support. Our minimum Gtk
dependency of >= 3.10 depends on Pango >= 1.32.4. Therefore via
Gtk, we depend on a plenty high enough version automatically. This
patch removes the explicit check. It also removes the single set
of guards for this, though it's effectively commented out anyway.