meson: Update plugins option to reflect what it now does
Previously, the 'plugins' option would toggle libpurple plugin
support. Now that libpurple's plugin support is required, the
'plugins' option instead merely toggles building plugins
(including prpls). Ultimately we probably want an option for
toggling the building of plugins and a separate one for prpls,
but that can come later. This patch updates the description of
the 'plugins' option to better reflect what it now does.