libpurple: Remove PURPLE_PLUGINS define from purple.h
The PURPLE_PLUGINS definition in purple.h is no longer needed now
that nothing uses it. This patch removes it, the PLUGINS_DEFINE
configuration data which was used to add it, and moves the
purple_h_conf configuration data variable into libpurple/meson.build
as nothing else in the root meson.build file uses it.